+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to QETLAB will be documented in this file.
 - EntangledSubspace: Constructs a basis of a bipartite r-entangled subspace of any dimension.
 - EntFormation: Computes the entanglement of formation of a 2-qubit state or a pure state.
 - MatsumotoFidelity: Computes the Matsumoto fidelity of two density matrices.
+- RandomPPTState: Generates a random density matrix with positive partial transpose, and optionally low rank.
 - helpers/asymind: Creates all vectors with strictly increasing permutations of an input vector.
 - helpers/dec_to_bin: Converts a decimal number to a binary vector. Replaces de2bi from the Communications toolbox.
 - helpers/pad_array: Pads an array with zeroes. Replaces padarray from the Image Processing toolbox.